+++ lib/lite/uxid/record/obfuscateid.rb
@@ -17,7 +17,7 @@ module Lite
         module ClassMethods
 
           def find_by_uxid(uxid)
-            decoded_id = Lite::Uxid::Reversible::Obfuscateid.decode(uxid, prefix: new.uxid_prefix)
+            decoded_id = Lite::Uxid::Reversible::Obfuscateid.decode(uxid)
             find_by(id: decoded_id)
           end
 
@@ -33,17 +33,13 @@ module Lite
         def id_to_uxid
           return unless respond_to?(:uxid)
 
-          Lite::Uxid::Reversible::Obfuscateid.encode(id, prefix: uxid_prefix)
+          Lite::Uxid::Reversible::Obfuscateid.encode(id)
         end
 
         def uxid_to_id
           return unless respond_to?(:uxid)
 
-          Lite::Uxid::Reversible::Obfuscateid.decode(uxid, prefix: uxid_prefix)
-        end
-
-        def uxid_prefix
-          nil
+          Lite::Uxid::Reversible::Obfuscateid.decode(uxid)
         end
 
         private