+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.1] - 2024-09-23
 ### Changed
 - Renamed `Scatterid` to `Obfuscateid`
+### Removed
+- Removed prefix option from obfuscateid
 
 ## [2.0.0] - 2024-09-23
 ### Added